Password manager
This passoword manager allows the user to save the email and a password for a website. The user can create their own password or click on the Password Generator and have a secure password created for them. Then the app creates a json object which saves all the passwords. The user can also look up a saved password by searching for the website.