Flash cards app for language learning
The flash cards app uses the TK Inter library to show flashcards to the user.
The current language is german.
The app shows the user a german word and the card flips after 3 seconds to show the english meaning.
The user can press the green button meaning they know the meaning or the red button.
If the green button is pressed then the word is removed from the list of words to be learned.